requirements.txt
wheel
Scrapy==2.6.2
scrapy-splash==0.8.0
scrapy-user-agents==0.1.1

python-dotenv==0.19.2
python-decouple==3.0

lxml==4.9.1
requests==2.28.1

pyopenssl==22.0.0
cryptography==36.0.0
dateparser==1.1.2
unidecode==1.3.4

psycopg2-binary==2.9.5
random_user_agent==1.0.1

jsonlines==3.1.0
google-cloud-storage